Move logic to App component
Move logic to App component.
