+++ src/components/Todo.jsx
@@ -1,15 +1,13 @@
 import CompleteButton from './CompleteButton';
 
-export default function Todo({ todo, setTodos }) {
+export default function Todo({ todo, handleDelete }) {
   const { id, value } = todo;
+
   return (
     <>
       <li key={id}>
         {value}
-        <CompleteButton
-          todoId={id}
-          setTodos={setTodos}
-        />
+        <CompleteButton handleDelete={handleDelete} />
       </li>
     </>
   );
