+++ src/components/TodoForm.jsx
@@ -1,31 +1,4 @@
-import { useState } from 'react';
-
-function newId(list) {
-  if (!list.length) {
-    return 1;
-  }
-
-  const id = list[list.length - 1].id + 1;
-
-  return id;
-}
-
-export default function TodoForm({ setTodos }) {
-  const [input, updateInput] = useState('');
-
-  function handleSubmit(e) {
-    e.preventDefault();
-
-    setTodos((todos) => [...todos, { id: newId(todos), value: input }]);
-    updateInput('');
-  }
-
-  function handleChange(e) {
-    const { value } = e.target;
-
-    updateInput(value);
-  }
-
+export default function TodoForm({ input, handleSubmit, handleChange }) {
   return (
     <form onSubmit={handleSubmit}>
       <input
