--- src/components/Todos.jsx
@@ -1,11 +1,0 @@
-import Todo from './Todo';
-
-export default function Todos({ todos, setTodos }) {
-  return (
-    <ol>
-      {todos.map((todo) => (
-        <Todo todo={todo} setTodos={setTodos} />
-      ))}
-    </ol>
-  );
-}
